moved icons
