+++ frontend/src/atoms/Icons/ClientsIcon.tsx
