+++ frontend/src/atoms/Icons/Ellipses.tsx
