+++ frontend/src/atoms/Icons/Logo.tsx
