+++ frontend/src/atoms/Icons/LogoClient.tsx
