+++ frontend/src/atoms/Icons/LogoEntreprise.tsx
