+++ frontend/src/atoms/Icons/LogoMark.tsx
