+++ frontend/src/atoms/Icons/ProjectsIcon.tsx
