+++ frontend/src/components/Navigation.tsx
@@ -1,12 +1,12 @@
 import { NavLink } from "react-router-dom";
-import DashboardIcon from "./Icons/DashboardIcon";
-import ProjectsIcon from "./Icons/ProjectsIcon";
-import PlusIcon from "./Icons/PlusIcon";
-import ClientsIcon from "./Icons/ClientsIcon";
-import SettingsIcon from "./Icons/SettingsIcon";
-import LogoEntrepriseIcon from "./Icons/LogoEntreprise";
+import DashboardIcon from "../atoms/Icons/DashboardIcon";
+import ProjectsIcon from "../atoms/Icons/ProjectsIcon";
+import PlusIcon from "../atoms/Icons/PlusIcon";
+import ClientsIcon from "../atoms/Icons/ClientsIcon";
+import SettingsIcon from "../atoms/Icons/SettingsIcon";
+import LogoEntrepriseIcon from "../atoms/Icons/LogoEntreprise";
 import { useRoleTheme } from "../context/roleThemeContext";
-import LogoClientIcon from "./Icons/LogoClient";
+import LogoClientIcon from "../atoms/Icons/LogoClient";
 
 export default function Navigation() {
   const role = useRoleTheme();
